chore: Fix main file, update code to utils
