+++ evaluations/evaluations.txt
@@ -1,3 +1,3 @@
-depression_severity: 5
-symptom_severity: 5
+depression_severity: 4
+symptom_severity: 4
 cognitive_distortion: 5
